alloc code done

+++ MP4/MP4_Sources/MP4_Sources/vm_pool.C
@@ -131,7 +131,7 @@ void VMPool::release(unsigned long _start_address) {
 
 bool VMPool::is_legitimate(unsigned long _address) {
 
-                Console::puts("         -> is_legitimate: base_address     =");Console::puti(this);Console::puts("\n");
+                Console::puts("         -> is_legitimate: base_address     =");Console::puti((unsigned int) this);Console::puts("\n");
                 Console::puts("         -> is_legitimate: checking address =");Console::puti(_address);Console::puts("\n");
 
     // if address belongs to free/alloc lists' region
